Correct parameter type in booking DAO method

+++ src/main/java/com/ss/training/utopia/counter/dao/BookingDao.java
@@ -14,6 +14,6 @@ import com.ss.training.utopia.counter.entity.BookingPk;
 public interface BookingDao extends JpaRepository<Booking, BookingPk> {
 
 	@Query("SELECT b FROM tbl_booking b JOIN tbl_flight f WHERE b.travelerId = ?1 AND b.active = true AND f.departTime > CURRENT_TIMESTAMP")
-	public List<Booking> findCancellable(Integer travelerId);
+	public List<Booking> findCancellable(Long travelerId);
 
 }
